requirements.txt
numpy==1.16.3
tensorflow==1.13.1
Keras==2.2.4
opencv-python==4.1.0.25
opencv-contrib-python==4.1.0.25
mtcnn==0.0.9
Pillow==6.0.0
scipy==1.3.0
scikit-learn==0.21.3